Nicer colors

+++ src/com/rojel/fractals/Plottable.java
@@ -1,5 +1,5 @@
 package com.rojel.fractals;
 
 public interface Plottable {
-	public float getPixel(double x, double y, double xRes, double yRes);
+	public int getPixel(double x, double y, double xRes, double yRes);
 }
