+++ src/com/rojel/fractals/plottables/Parabola.java
@@ -4,7 +4,7 @@ import com.rojel.fractals.Plottable;
 
 public class Parabola implements Plottable {
 	@Override
-	public float getPixel(double x, double y, double xRes, double yRes) {
+	public int getPixel(double x, double y, double xRes, double yRes) {
 		if (Math.abs(y - function(x)) < 2 * yRes)
 			return 1;
